refactor: streamline deployment logic and enhance wallet handling across node and web environments
