+++ src/lib/deploy/deploy.common.ts
@@ -1,4 +1,4 @@
-import { connect } from "@permaweb/aoconnect";
+import { connect, createDataItemSigner } from "@permaweb/aoconnect";
 import pLimit from "p-limit";
 import type {
   AosConfig,
@@ -6,13 +6,22 @@ import type {
   DeployResult,
   Services
 } from "../../types";
-import { AOS_QUERY, defaultServices } from "../constants";
+import { AOS_QUERY, APP_NAME, defaultServices } from "../constants";
+import { Logger } from "../logger";
+import { minifyLuaCode } from "../minify";
 import {
   getArweave,
+  hasValidBlueprints,
+  isArweaveAddress,
   isCronPattern,
   isUrl,
+  loadBlueprints,
+  logActionStatus,
+  parseToInt,
+  pollForProcessSpawn,
   retryWithDelay
 } from "../utils/utils.common";
+import { WalletInterface } from "../wallet/wallet.types";
 
 /**
  * Manages deployments of contracts to AO.
@@ -109,6 +118,207 @@ export class BaseDeploymentsManager {
     if (!isCronValid) {
       throw new Error("Invalid cron flag!");
     }
+  }
+
+  protected async _deployContract({
+    name,
+    wallet: walletInstance,
+    contractPath,
+    getContractSource,
+    tags,
+    cron,
+    module,
+    scheduler,
+    retry,
+    configName,
+    processId,
+    sqlite,
+    services,
+    minify,
+    contractTransformer,
+    onBoot,
+    blueprints,
+    silent = false,
+    forceSpawn = false
+  }: Omit<DeployConfig, "wallet"> & {
+    getContractSource: () => Promise<string>;
+    wallet: WalletInterface;
+  }): Promise<DeployResult> {
+    name = name || "default";
+    configName = configName || name;
+    retry = {
+      count: parseToInt(retry?.count, 10),
+      delay: parseToInt(retry?.delay, 3000)
+    };
+
+    const logger = new Logger(configName, silent);
+    const aosConfig = await this.getAosConfig();
+    module = isArweaveAddress(module)
+      ? module!
+      : sqlite
+        ? aosConfig.sqliteModule
+        : aosConfig.module;
+    scheduler = isArweaveAddress(scheduler) ? scheduler! : aosConfig.scheduler;
+
+    const owner = await walletInstance.getAddress();
+
+    const signer = createDataItemSigner(walletInstance.signer);
+    services = this.validateServices(services);
+
+    // Initialize the AO instance with validated URLs
+    const aoInstance = this.getAoInstance(services);
+
+    logActionStatus(
+      "deploy",
+      logger,
+      contractPath || "provided contract source",
+      blueprints
+    );
+
+    let isNewProcess = forceSpawn;
+
+    if (
+      !forceSpawn &&
+      (!processId || (processId && !isArweaveAddress(processId)))
+    ) {
+      processId = await this.findProcess(
+        name,
+        owner,
+        retry,
+        services.gatewayUrl!
+      );
+      isNewProcess = !processId;
+    }
+
+    const contractSrc = await getContractSource();
+
+    let contractSource = contractSrc || "";
+    let blueprintsSrc = "";
+
+    if (!contractSource && !hasValidBlueprints(blueprints)) {
+      throw new Error(
+        "Please provide either contract source code or blueprints."
+      );
+    }
+
+    if (Array.isArray(blueprints) && blueprints.length > 0) {
+      blueprintsSrc = await loadBlueprints(blueprints);
+    }
+
+    if (blueprintsSrc || contractSource) {
+      contractSource = [blueprintsSrc, contractSource]
+        .filter(Boolean)
+        .join("\n\n")
+        .trim();
+    }
+
+    if (contractTransformer && typeof contractTransformer === "function") {
+      logger.log("Transforming contract...", false, false);
+      contractSource = await contractTransformer(contractSource);
+    }
+
+    if (minify) {
+      logger.log("Minifying contract...", false, false);
+      contractSource = await minifyLuaCode(contractSource);
+    }
+
+    if (isNewProcess) {
+      logger.log("Spawning new process...", false, true);
+      tags = Array.isArray(tags) ? tags : [];
+      tags = [
+        { name: "App-Name", value: APP_NAME },
+        { name: "Name", value: name },
+        { name: "aos-Version", value: "REPLACE-AO-DEPLOY-VERSION" },
+        { name: "Authority", value: aosConfig.authority },
+        ...tags
+      ];
+
+      if (onBoot) {
+        tags = [...tags, { name: "On-Boot", value: "Data" }];
+      }
+
+      if (cron) {
+        this.validateCron(cron);
+        tags = [
+          ...tags,
+          { name: "Cron-Interval", value: cron },
+          { name: "Cron-Tag-Action", value: "Cron" }
+        ];
+      }
+
+      const data = onBoot ? contractSource : "1984";
+      processId = await retryWithDelay(
+        () => aoInstance.spawn({ module, signer, tags, data, scheduler }),
+        retry.count,
+        retry.delay
+      );
+
+      if (!processId) {
+        throw new Error("Failed to spawn process");
+      }
+
+      await pollForProcessSpawn({
+        processId: processId!,
+        gatewayUrl: services.gatewayUrl
+      });
+
+      if (onBoot) {
+        return { name, processId, isNewProcess, configName };
+      }
+    }
+
+    let messageId: string;
+    if (!onBoot || !isNewProcess) {
+      if (!isNewProcess) {
+        logger.log("Updating existing process...", false, true);
+      }
+      // Load contract to process
+      messageId = await retryWithDelay(
+        async () =>
+          aoInstance.message({
+            process: processId!,
+            tags: [{ name: "Action", value: "Eval" }],
+            data: contractSource,
+            signer
+          }),
+        retry.count,
+        retry.delay
+      );
+
+      const { Output, Error: error } = await retryWithDelay(
+        async () =>
+          aoInstance.result({
+            process: processId!,
+            message: messageId
+          }),
+        retry.count,
+        retry.delay
+      );
+
+      let errorMessage = null;
+
+      if (Output?.data?.output) {
+        errorMessage = Output.data.output;
+      } else if (error) {
+        if (typeof error === "object" && Object.keys(error).length > 0) {
+          errorMessage = JSON.stringify(error);
+        } else {
+          errorMessage = String(error);
+        }
+      }
+
+      if (errorMessage) {
+        throw new Error(errorMessage);
+      }
+    }
+
+    return {
+      name,
+      processId: processId!,
+      messageId: messageId!,
+      isNewProcess,
+      configName
+    };
   }
 
   /**
