+++ src/lib/deploy/deploy.web.ts
@@ -1,17 +1,5 @@
-import { createDataItemSigner } from "@permaweb/aoconnect";
-import { JWKInterface } from "arweave/web/lib/wallet";
-import type { DeployConfig, DeployResult } from "../../types";
-import { APP_NAME } from "../constants";
-import { Logger } from "../logger";
-import {
-  hasValidBlueprints,
-  isArweaveAddress,
-  loadBlueprints,
-  logActionStatus,
-  parseToInt,
-  pollForProcessSpawn,
-  retryWithDelay
-} from "../utils/utils.common";
+import pLimit from "p-limit";
+import type { DeployResult, WebDeployConfig } from "../../types";
 import { Wallet } from "../wallet/wallet.web";
 import { BaseDeploymentsManager } from "./deploy.common";
 
@@ -21,203 +9,47 @@ import { BaseDeploymentsManager } from "./deploy.common";
 export class DeploymentsManager extends BaseDeploymentsManager {
   /**
    * Deploys or updates a contract on AO.
-   * @param {DeployConfig} deployConfig - Configuration options for the deployment.
+   * @param {WebDeployConfig} deployConfig - Configuration options for the deployment.
    * @returns {Promise<DeployResult>} The result of the deployment.
    */
-  async deployContract({
-    name,
-    wallet,
-    contractSrc,
-    tags,
-    cron,
-    module,
-    scheduler,
-    retry,
-    configName,
-    processId,
-    sqlite,
-    services,
-    contractTransformer,
-    onBoot,
-    blueprints,
-    silent = false,
-    forceSpawn = false
-  }: DeployConfig): Promise<DeployResult> {
-    name = name || "default";
-    configName = configName || name;
-    retry = {
-      count: parseToInt(retry?.count, 10),
-      delay: parseToInt(retry?.delay, 3000)
-    };
+  async deployContract(deployConfig: WebDeployConfig): Promise<DeployResult> {
+    const walletInstance = await Wallet.load();
 
-    const logger = new Logger(configName, silent);
-    const aosConfig = await this.getAosConfig();
-    module = isArweaveAddress(module)
-      ? module!
-      : sqlite
-        ? aosConfig.sqliteModule
-        : aosConfig.module;
-    scheduler = isArweaveAddress(scheduler) ? scheduler! : aosConfig.scheduler;
+    const getContractSource = async () => deployConfig.contractSrc || "";
 
-    const walletInstance = await Wallet.load(wallet as JWKInterface);
-    const owner = await walletInstance.getAddress();
-    const signer = createDataItemSigner(
-      walletInstance.jwk || window.arweaveWallet
+    return this._deployContract({
+      ...deployConfig,
+      wallet: walletInstance,
+      getContractSource
+    });
+  }
+
+  /**
+   * Deploys multiple contracts concurrently with specified concurrency limits.
+   * @param {WebDeployConfig[]} deployConfigs - Array of deployment configurations.
+   * @param {number} concurrency - Maximum number of deployments to run concurrently. Default is 5.
+   * @returns {Promise<PromiseSettledResult<DeployResult>[]>} Array of results for each deployment, either fulfilled or rejected.
+   */
+  async deployContracts(
+    deployConfigs: WebDeployConfig[],
+    concurrency: number = 5
+  ): Promise<PromiseSettledResult<DeployResult>[]> {
+    const limit = pLimit(concurrency);
+    const promises = deployConfigs.map((config) =>
+      limit(() => this.deployContract(config))
     );
-    services = this.validateServices(services);
-
-    // Initialize the AO instance with validated URLs
-    const aoInstance = this.getAoInstance(services);
-
-    logActionStatus("deploy", logger, "provided contract source", blueprints);
-
-    let isNewProcess = forceSpawn;
-
-    if (
-      !forceSpawn &&
-      (!processId || (processId && !isArweaveAddress(processId)))
-    ) {
-      processId = await this.findProcess(
-        name,
-        owner,
-        retry,
-        services.gatewayUrl!
-      );
-      isNewProcess = !processId;
-    }
-
-    let contractSource = contractSrc || "";
-    let blueprintsSrc = "";
-
-    if (!contractSource && !hasValidBlueprints(blueprints)) {
-      throw new Error(
-        "Please provide either contract source code or blueprints."
-      );
-    }
-
-    if (Array.isArray(blueprints) && blueprints.length > 0) {
-      blueprintsSrc = await loadBlueprints(blueprints);
-    }
-
-    if (blueprintsSrc || contractSource) {
-      contractSource = [blueprintsSrc, contractSource]
-        .filter(Boolean)
-        .join("\n\n")
-        .trim();
-    }
-
-    if (contractTransformer && typeof contractTransformer === "function") {
-      logger.log("Transforming contract...", false, false);
-      contractSource = await contractTransformer(contractSource);
-    }
-
-    if (isNewProcess) {
-      logger.log("Spawning new process...", false, true);
-      tags = Array.isArray(tags) ? tags : [];
-      tags = [
-        { name: "App-Name", value: APP_NAME },
-        { name: "Name", value: name },
-        { name: "aos-Version", value: "REPLACE-AO-DEPLOY-VERSION" },
-        { name: "Authority", value: aosConfig.authority },
-        ...tags
-      ];
-
-      if (onBoot) {
-        tags = [...tags, { name: "On-Boot", value: "Data" }];
-      }
-
-      if (cron) {
-        this.validateCron(cron);
-        tags = [
-          ...tags,
-          { name: "Cron-Interval", value: cron },
-          { name: "Cron-Tag-Action", value: "Cron" }
-        ];
-      }
-
-      const data = onBoot ? contractSource : "1984";
-      processId = await retryWithDelay(
-        () => aoInstance.spawn({ module, signer, tags, data, scheduler }),
-        retry.count,
-        retry.delay
-      );
-
-      if (!processId) {
-        throw new Error("Failed to spawn process");
-      }
-
-      await pollForProcessSpawn({
-        processId: processId!,
-        gatewayUrl: services.gatewayUrl
-      });
-
-      if (onBoot) {
-        return { name, processId, isNewProcess, configName };
-      }
-    }
-
-    let messageId: string;
-    if (!onBoot || !isNewProcess) {
-      if (!isNewProcess) {
-        logger.log("Updating existing process...", false, true);
-      }
-      // Load contract to process
-      messageId = await retryWithDelay(
-        async () =>
-          aoInstance.message({
-            process: processId!,
-            tags: [{ name: "Action", value: "Eval" }],
-            data: contractSource,
-            signer
-          }),
-        retry.count,
-        retry.delay
-      );
-
-      const { Output, Error: error } = await retryWithDelay(
-        async () =>
-          aoInstance.result({
-            process: processId!,
-            message: messageId
-          }),
-        retry.count,
-        retry.delay
-      );
-
-      let errorMessage = null;
-
-      if (Output?.data?.output) {
-        errorMessage = Output.data.output;
-      } else if (error) {
-        if (typeof error === "object" && Object.keys(error).length > 0) {
-          errorMessage = JSON.stringify(error);
-        } else {
-          errorMessage = String(error);
-        }
-      }
-
-      if (errorMessage) {
-        throw new Error(errorMessage);
-      }
-    }
-
-    return {
-      name,
-      processId: processId!,
-      messageId: messageId!,
-      isNewProcess,
-      configName
-    };
+    const results = await Promise.allSettled(promises);
+    return results;
   }
 }
 
 /**
  * Deploys or updates a contract on AO.
- * @param {DeployConfig} deployConfig - Configuration options for the deployment.
+ * @param {WebDeployConfig} deployConfig - Configuration options for the deployment.
  * @returns {Promise<DeployResult>} The result of the deployment.
  */
 export async function deployContract(
-  deployConfig: DeployConfig
+  deployConfig: WebDeployConfig
 ): Promise<DeployResult> {
   const manager = new DeploymentsManager();
   return manager.deployContract(deployConfig);
@@ -225,12 +57,12 @@ export async function deployContract(
 
 /**
  * Deploys multiple contracts concurrently with specified concurrency limits.
- * @param {DeployConfig[]} deployConfigs - Array of deployment configurations.
+ * @param {WebDeployConfig[]} deployConfigs - Array of deployment configurations.
  * @param {number} concurrency - Maximum number of deployments to run concurrently. Default is 5.
  * @returns {Promise<PromiseSettledResult<DeployResult>[]>} Array of results for each deployment, either fulfilled or rejected.
  */
 export async function deployContracts(
-  deployConfigs: DeployConfig[],
+  deployConfigs: WebDeployConfig[],
   concurrency: number = 5
 ): Promise<PromiseSettledResult<DeployResult>[]> {
   const manager = new DeploymentsManager();
