+++ src/lib/wallet/wallet.types.ts
@@ -1,0 +1,6 @@
+import { JWKInterface } from "arweave/node/lib/wallet";
+
+export interface WalletInterface {
+  signer: JWKInterface | Window["arweaveWallet"];
+  getAddress: () => Promise<string>;
+}
